Upgrade to Terraform v0.11.11
This change upgrades to Terraform v0.11.11, the latest stable release.

In doing so, it also uses govendor-override to generate the Gopkg files.
I'm not sure, but I assume that the prior files were generated by hand
-- or an earlier version of the tool -- to resolve various conflicts
that led to an unbuildable vendor directory. This more completely
specifies those constraints/overrides, while also making it easy to
regenerate using the govendor-override tool.

This is part of my attempt at ingesting this repo as a dependency...