Eliminate non-determinitic visitation
This ensures that anytime we enumerate a map's keys/values when
visiting the IL AST, we do so in a deterministic order. Without
this, various elements of the output source changes from one run
to the next. I'm not claiming I've found all sources, but this
did address the `all().apply()` calls uncovered in pulumi/tf2pulumi#32.