Statically coerce string literals to numbers and bools (#57)
Add a target-independent pass to insert coercions. This pass statically
coerces strings to numbers or bools in line with the behavior of the TF
`helper/schema` package's field reader.

Also, improve numeric literal generation to use only as many significant
digits as are necessary to uniquely identify the number.

Fixes #55.