Add support for filtering resource names (#66)
Our providers commonly support auto-naming resources if a name is not
specified in order to help avoid name collisions. These changes add a
flag to `tf2pulumi` that can be used to specify the key of a name
property to filter from all resources in order to allow the autonamer to
go to work.

Fixes #43.