Fix data sources with output-typed inputs (#74)
Data sources with output-typed inputs need to be generated as a call to
`apply` that contains a call to the data source itself with the resolved
outputs. These changes implement this in a moderately subtle fashion:
instead of generating the data source ad-hoc, it is turned into an
intrinsic that is then handed over to the apply rewriter. If the apply
rewriter does not transform the intrinsic, the call to the data source
is wrapped in a call to `pulumi.output` in order to keep the generated
code in normal form.

Fixes #67.