Add skeleton for Python backend (#75)
* Add skeleton for Python backend

This commit adds a skeleton backend for Python that does nothing. It
hooks up the relevant plumbing bits in the frontend to select Python as
a target language and generate a Python generator if requested. The
backend basically does nothing and emits NYIs for everything.

Future commits will incrementally build up the Python backend.

* CR feedback: createGenerator -> newGenerator