Improve generation order.
Rather than generating nodes in order by kind and then alphabetically by
name, implement a new algorithm that attempts to keep definitions
grouped by source file.

This algorithm works by building a worklist of files to generate and
iterating until the worklist is empty. On each iteration, the nodes from
the file with the the fewest references to nodes that have not yet been
generated and that are defined in other files is chosen for generation.
Within each file, nodes are ordered by their appearence in the original
source file. Out-of-order generation only occurs when it is necessary to
satisfy the requirement that nodes are generated in a valid topological
order.