Add support for generating conditional resources. (#131)
* Add support for marking conditional resources.

These changes add a new helper to the `il` package that is able to find
resources with a count that maps to (0, 1). These resources can be
considered conditionally instantiated by downstream code generation.

* Add support for generating conditional resources.

Conditionally-instantiated resources are now generated using `if`
statements rather than `for` loops. For most resources, the generated
code is as simple as:

```
let foo: Foo | undefined;
if (condition) {
    foo = new Foo(...);
}
```

For resources with properties that reference the condition variable,
the code generation is a little bit more complicated. In this case, we
assign the result of the condition to a local so that the resource has
something to reference:

```
const createFoo = condition;
let foo: Foo | undefined;
if (createFoo) {
    foo = new Foo(...);
}
```

Fixes #123.