Upgrade dependencies.
Notabluy:
- terraform -> terraform v0.12.9
- pulumi-terraform -> pulumi-terraform-bridge

The first change required that we vendor in the TF11 module loader and
config package, as the new version of TF12 has removed that code.

The second change required that we link in a very slightly modified
version of terraform-provider-sdk, as that package registers types with
gob that conflict with types registered by terraform itself.

With these changes, we now have a setup that will allow us to update
tf2pulumi to understand both TF11 and TF12 config in the same binary.