Add preliminary HCL2 + Python support (#160)
These changes add preliminary support for HCL2 and Python to
`tf2pulumi`. HCL2 support is used if it is explicitly requested or if
the input configuration cannot be treated as HCL1.

As part of these changes, `tf2pulumi` no longer depends on the
Terraform repository, and instead vendors in the slice it needs.
This involves one compromise: TF11 modules that are sourced
from a registry are officially unsupported. Prior to these changes
these modules were only unofficially unsupported.